use non-: version of custom data
PhysiCell defaults to using a : in custom data signals/behaviors.
This interferes with pcvct using : for finding attributes in an XML path.
Now, when reading in a CSV, make sure custom signals/behaviors are stored in the XML with the form "custom <name>".
Then, when writing out, write to the form "custom:<name>".